+++ src/main/java/com/project/shop/controller/admin/AdminItemsController.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class adminItemsController {
+public class AdminItemsController {
 
 	@GetMapping(value = "/admin/ItemList")
 	   public String adminItemsList() {
