+++ src/main/java/com/project/shop/controller/admin/AdminMemberController.java
@@ -4,9 +4,10 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class adminOrderController {
-   @GetMapping(value = "/admin/adminOrder")
-   public String adminOrder() {
-	   return "admin/admin_order";
-   }
+public class AdminMemberController {
+
+	@GetMapping(value = "/admin/member")
+	   public String adminItemsList() {
+		   return "admin/admin_member";
+	   }
 }
