+++ src/main/java/com/project/shop/controller/items/ItemsController.java
@@ -1,0 +1,13 @@
+package com.project.shop.controller.items;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class ItemsController {
+
+	@GetMapping(value = "/items")
+	   public String items() {
+		   return "items/items";
+	   }
+}
