+++ src/main/resources/static/css/admin_member.css
@@ -1,0 +1,91 @@
+@import url(//fonts.googleapis.com/earlyaccess/notosanskr.css);
+
+.notosanskr * { 
+ font-family: 'Noto Sans KR', sans-serif;
+}
+
+.subtitle_main .find {
+    height: 33px;
+    width: 400px;
+    border: 1px solid #1b5ac2;
+    background: #ffffff;
+    margin-left: 400px;
+    margin-top: 50px;
+}
+
+.subtitle_main .find input {
+    font-size: 16px;
+    width: 325px;
+    padding: 10px;
+    border: 0px;
+    outline: none;
+    float: left;
+}
+
+.subtitle_main .find button {
+    width: 50px;
+    height: 100%;
+    border: 0px;
+    background: #1b5ac2;
+    outline: none;
+    float: right;
+    color: #ffffff;
+}
+
+.subtitle_main .memberlist_graph {
+    border: 1px solid black;
+    height: 20%; width: 66%;
+    margin-top: 50px;
+    margin-left: 17%;
+    margin-bottom: 50px;
+    font-family: 'Noto Sans KR', sans-serif;
+}
+
+.subtitle_main .memberlist_graph .memberlist_graphtable {
+    width: 100%;
+    border: 1px solid red;   
+}
+
+.subtitle_main .memberlist_graph .memberlist_graphtable .memberlist_graphtable_title {
+    text-align: center;
+    background-color: rgb(234,234,234);
+    font-weight: bold;
+}
+
+.subtitle_main .memberlist_graph .memberlist_graphtable tr td {
+    border: 1px solid black;
+    height: 45px;
+}
+
+.subtitle_main .memberlist_graph .memberlist_graphtable tr td:nth-child(1) {
+    width: 40%;    
+    text-align: center;
+}
+
+.subtitle_main .memberlist_graph .memberlist_graphtable tr td:nth-child(2) {
+    width: 30%;    
+    text-align: center;
+}
+
+.subtitle_main .memberlist_graph .memberlist_graphtable tr td:nth-child(3) {
+    width: 20%;    
+    text-align: center;
+}
+
+.subtitle_main .memberlist_graph .memberlist_graphtable tr td:nth-child(4) {
+    width: 10%;    
+    text-align: center;
+}
+
+.subtitle_main .memberlist_graph .memberlist_graphtable tr td:nth-child(4) button {
+    background-color: rgb(52,152,219);
+    color: white;
+    border: 0px solid red;
+    border-radius: 5px;
+    width: 50px;
+    height: 30px;
+}
+
+.memberlist1 {
+    font-size: 14px;
+}
