+++ src/main/resources/templates/admin/admin_member.html
@@ -13,6 +13,77 @@
     <div class = "subtitle_title">
         <p>회원 관리</p>
     </div>   
+    
+    <!-- 이메일로 검색표시창 -->
+    <div class = "find">
+        <input type="text" style = "height:30px" placeholder="이메일을 입력해주세요">
+        <button>
+            검색
+        </button>
+    </div>
+
+    <!-- 상품 수정/삭제 표-->
+    <div class = "memberlist_graph">
+        <table class = "memberlist_graphtable">
+            <tr class = "memberlist_graphtable_title">
+                <td>이메일</td>
+                <td>전화번호</td>
+                <td>가입날짜</td>
+                <td>탈퇴</td>
+            </tr>
+            <tr class = "memberlist1">
+                <td>tlsehddms1234@naver.com</td>
+                <td>010-5646-1998</td>
+                <td>2018.01.01</td>
+                <td>
+                    <button type = "button">탈퇴</button>
+                </td>
+            </tr>
+            <tr class = "memberlist1">
+                <td></td>
+                <td></td>
+                <td></td>
+                <td></td>
+            </tr>
+            <tr class = "memberlist1">
+                <td></td>
+                <td></td>
+                <td></td>
+                <td></td>
+            </tr>
+            <tr class = "memberlist1">
+                <td></td>
+                <td></td>
+                <td></td>
+                <td></td>
+            </tr>
+            <tr class = "memberlist1">
+                <td></td>
+                <td></td>
+                <td></td>
+                <td></td>
+            </tr>
+        </table>
+    </div>
+
+     <!-- 밑에 숫자 넘어가기 -->
+     <nav aria-label="Page navigation example">
+        <ul class="pagination justify-content-center">
+          <li class="page-item">
+            <a class="page-link" href="#" aria-label="Previous">
+              <span aria-hidden="true">&laquo;</span>
+            </a>
+          </li>
+          <li class="page-item"><a class="page-link" href="#">1</a></li>
+          <li class="page-item"><a class="page-link" href="#">2</a></li>
+          <li class="page-item"><a class="page-link" href="#">3</a></li>
+          <li class="page-item">
+            <a class="page-link" href="#" aria-label="Next">
+              <span aria-hidden="true">&raquo;</span>
+            </a>
+          </li>
+        </ul>
+      </nav>
 
     
 </article>
